Return Rentals + Max Rentals
You can now return your rental for a set price I want to make it so in the future it will be a fraction of the price to rent. There is a set max of rentals now you can change in the config. THIS MAKES THE RESOURCE LESS DRAG AND DROP NOW YOU WILL HAVE TO EDIT CLIENT.LUA.